% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasCorrectionIG.R
\name{BiasCorrectionIG}
\alias{BiasCorrectionIG}
\title{BiasCorrectionIG}
\usage{
BiasCorrectionIG(D0, nf, ncl, dist, para, nMonte)
}
\arguments{
\item{nf}{The number of covariates under study.}

\item{ncl}{The number of clusters.}

\item{para}{The vector of the ML estimates of unknown parameters. This vector has a form
\deqn{y = (ln(a), \beta , ln(\sigma ^2))}
for the exponential baseline hazard function,
\deqn{y = (ln(a), ln(b), \beta , ln(\sigma ^2))}
for the Weibull baseline hazard function and
\deqn{y = (ln(10^3a), ln(10b), \beta , ln(\sigma ^2))}
for the Gompertz baseline hazard function, where a and b are the slope and the shape parameters,
\eqn{\beta } are the Cox-regression parameters, and \eqn{\sigma ^2} is the variance of frailty.}

\item{nMonte}{The number of the resamplings for the Monte Carlo integrations}

\item{D}{The data frame. The first 'nf' columns are the covariates under study. Other columns must include fields
'time' for time-to-failure, 'Cens' for censoring and 'cluster' for cluster identification. The number of rows is equal to the number of subjects.}
}
\value{
Two fields - the bias correction \eqn{Bias} and its history \eqn{Biasdiag} for ML estimates
(for \eqn{nMonte} replacements) in the form:
\deqn{y1 = (ln(a), \beta , ln(\sigma ^2))}
for the exponential baseline hazard function,
\deqn{y1 = (ln(a), \beta , ln(\sigma ^2), b)}
for the Weibull baseline hazard function and
\deqn{y1 = (ln(10^3a), \beta , ln(\sigma ^2), 10b)}
If the calculation of the bias correction is not possible returns NAs.
}
\description{
Returns the bias correction for ML estimates in the model with inverse gaussian frailty.
}
\details{
This function calculates the bias correction for the proportional hazards model with gamma frailty
in accordance with the Cox-Snell method.  The baseline cumulative hazard functions are:
\enumerate{
\item Exponential with baseline cumulative hazard function \deqn{H_{0}(t;a)=at;}
\item Weibull with baseline cumulative hazard function \deqn{H_{0}(t;a,b)=at^b;}
\item Gompertz with baseline cumulative hazard function \deqn{H_{0}(t;a,b)=a(exp(bt)-1)).}
}
}
\examples{
\dontrun{
BiasCorrectionIG(D,nf,ncl,para,nMonte)

}
}
